+++ ngs_mapper/runsample.py
@@ -227,17 +227,26 @@ def main():
         sys.exit(1)
     # So we can set the global logger
     global logger
+    # Setup analysis directory
     if os.path.isdir( args.outdir ):
         if os.listdir( args.outdir ):
             raise AlreadyExists( "{0} already exists and is not empty".format(args.outdir) )
-    else: os.mkdir(args.outdir)
-    tmpdir = args.outdir or os.environ.get('TMPDIR', tempfile.tempdir)
-    print "outdir %s, tmpdir %s" % (args.outdir, tmpdir)
+    else:
+        os.makedirs(args.outdir)
+
+    # tempdir root will be TMPDIR environ variable if it exists
+    # unless outdir is set
+    # allows user to specify TMPDIR somewhere else if they want such as
+    # /dev/shm
+    tmpdir = args.outdir
+    # Directory analysis is run in will be inside of tmpdir
     tdir = tempfile.mkdtemp('runsample', args.prefix, dir=tmpdir)
+    os.environ['TMPDIR'] = tdir
+
     bamfile = os.path.join( tdir, args.prefix + '.bam' )
     flagstats = os.path.join( tdir, 'flagstats.txt' )
-    consensus = os.path.join( tdir, bamfile+'.consensus.fasta' )
-    vcf = os.path.join( tdir, bamfile+'.vcf' )
+    consensus = bamfile+'.consensus.fasta'
+    vcf = bamfile+'.vcf'
     bwalog = os.path.join( tdir, 'bwa.log' )
     stdlog = os.path.join( tdir, args.prefix + '.std.log' )
     logfile = os.path.join( tdir, args.prefix + '.log' )
@@ -418,9 +427,12 @@ def pbs_job(runsampleargs, pbsargs, tmpdir=None):
         template += '#PBS -m abe\n' \
             '#PBS -M ' + qsub_args.qsub_M + '\n'
 
+<<<<<<< HEAD
     if tmpdir is not None:
         template += 'export TMPDIR=' + tmpdir
 
+=======
+>>>>>>> 06ec38b1d4ac09d486a92afa37448b5de092009e
     template += '\n' \
         'cd $PBS_O_WORKDIR\n' \
         'runsample {runsampleargs}\n'